+++ example/quant_analysis/pytorch_yolov6/run.sh
@@ -1,1 +1,3 @@
-python run.py --config_path=./config.yaml 
+export PYTHONPATH=/paddle/xuchang/Quant/PaddleSlim/:$PYTHONPATH
+#ps -aux | grep run | awk '{print $2}' | uniq | xargs kill -9
+python3.7 run.py --config_path=./config.yaml
